@@ -13,6 +13,8 @@ internal class CloudService
     [VectorStoreData]
     public string Description { get; set; }
 
-    [VectorStoreVector(Dimensions: 384, DistanceFunction = DistanceFunction.CosineSimilarity)]
+    [VectorStoreVector(
+        Dimensions: 384,
+        DistanceFunction = DistanceFunction.CosineSimilarity)]
     public ReadOnlyMemory<float> Vector { get; set; }
 }
